Move sample code in separate file; fix docs for merge()
