+++ bach/bach/dataframe.py
@@ -363,7 +363,7 @@ class DataFrame:
             order by ordinal_position;
         """
         with engine.connect() as conn:
-            sql = _escape_parameter_characters(conn, sql)
+            sql = escape_parameter_characters(conn, sql)
             res = conn.execute(sql)
         return {x[0]: get_dtype_from_db_dtype(x[1]) for x in res.fetchall()}
 
@@ -613,54 +613,14 @@ class DataFrame:
             All data in the DataFrame to be sampled is queried to create the sample.
         """
         # todo if_exists and overwrite are two different syntax for the same thing. should we align?
-        if sample_percentage is None and filter is None:
-            raise ValueError('Either sample_percentage or filter must be set')
-
-        original_node = self.get_current_node(name='before_sampling')
-        df = self
-        if filter is not None:
-            sample_percentage = None
-            from bach.series import SeriesBoolean
-            if not isinstance(filter, SeriesBoolean):
-                raise TypeError('Filter parameter needs to be a SeriesBoolean instance.')
-            # Boolean argument implies return type of self[filter], help mypy a bit
-            df = cast('DataFrame', self[filter])
-
-        if df._group_by is not None:
-            raise NotImplementedError('Dataframes that have an active grouping can currently not be sampled. '
-                                      'Call df.materialize() first.')
-
-        with df.engine.connect() as conn:
-            if overwrite:
-                sql = f'DROP TABLE IF EXISTS {quote_identifier(table_name)}'
-                sql = _escape_parameter_characters(conn, sql)
-                conn.execute(sql)
-
-            if sample_percentage:
-                repeatable = f'repeatable ({seed})' if seed else ''
-
-                sql = f'''
-                    create temporary table tmp_table_name on commit drop as
-                    ({df.view_sql()});
-                    create temporary table {quote_identifier(table_name)} as
-                    (select * from tmp_table_name
-                    tablesample bernoulli({sample_percentage}) {repeatable})
-                '''
-            else:
-                sql = f'create temporary table {quote_identifier(table_name)} as ({df.view_sql()})'
-            sql = _escape_parameter_characters(conn, sql)
-            conn.execute(sql)
-
-        # Use SampleSqlModel, that way we can keep track of the current_node and undo this sampling
-        # in get_unsampled() by switching this new node for the old node again.
-        new_base_node = SampleSqlModel(table_name=table_name, previous=original_node)
-
-        return DataFrame.get_instance(
-            engine=df.engine,
-            base_node=new_base_node,
-            index_dtypes=df.index_dtypes,
-            dtypes=df.dtypes,
-            group_by=None
+        from bach.sample import get_sample
+        return get_sample(
+            df=self,
+            table_name=table_name,
+            filter=filter,
+            sample_percentage=sample_percentage,
+            overwrite=overwrite,
+            seed=seed
         )
 
     def get_unsampled(self) -> 'DataFrame':
@@ -676,29 +636,8 @@ class DataFrame:
 
         :returns: an unsampled copy of the current sampled DataFrame.
         """
-        df = self
-        if df._group_by:
-            df = df.materialize(node_name='get_unsampled')
-
-        # The returned DataFrame has a modified base_node graph, in which the node that introduced the
-        # sampling is removed.
-        sampled_node_tuple = find_node(
-            start_node=df.base_node,
-            function=lambda node: isinstance(node, SampleSqlModel)
-        )
-        if sampled_node_tuple is None:
-            raise ValueError('No sampled node found. Cannot un-sample data that has not been sampled.')
-
-        assert isinstance(sampled_node_tuple.model, SampleSqlModel)  # help mypy
-        updated_graph = replace_node_in_graph(
-            start_node=df.base_node,
-            reference_path=sampled_node_tuple.reference_path,
-            replacement_model=sampled_node_tuple.model.previous
-        )
-
-        index = {s.name: s.copy_override(base_node=updated_graph) for s in df._index.values()}
-        series = {s.name: s.copy_override(base_node=updated_graph, index=index) for s in df._data.values()}
-        return df.copy_override(base_node=updated_graph, index=index, series=series)
+        from bach.sample import get_unsampled
+        return get_unsampled(df=self)
 
     def __getitem__(self,
                     key: Union[str, List[str], Set[str], slice, 'SeriesBoolean']) -> DataFrameOrSeries:
@@ -1359,7 +1298,7 @@ class DataFrame:
                      if series.dtype_to_pandas is not None}
 
             # read_sql_query expects a parameterized query, so we need to escape the parameter characters
-            sql = _escape_parameter_characters(conn, sql)
+            sql = escape_parameter_characters(conn, sql)
             pandas_df = pandas.read_sql_query(sql, conn).astype(dtype)
 
             if len(self._index):
@@ -1525,11 +1464,22 @@ class DataFrame:
         combined columns of both dataframes, and the rows that result from joining on the specified columns.
         The columns that are joined on can consist (partially or fully) out of index columns.
 
-        See :py:meth:`bach.merge.merge` for more information.
         The interface of this function is similar to pandas' merge, but the following parameters are not
         supported: `sort`, `copy`, `indicator`, and `validate`.
         Additionally when merging two frames that have conflicting columns names, and joining on indices,
         then the resulting columns/column names can differ slightly from Pandas.
+
+
+        :param right: DataFrame or Series to join on self
+        :param how: supported values: {‘left’, ‘right’, ‘outer’, ‘inner’, ‘cross’}
+        :param on: optional, column(s) to join left and right on.
+        :param left_on: optional, column(s) from the left df to join on
+        :param right_on: optional, column(s) from the right df/series to join on
+        :param left_index: If true uses the index of the left df as columns to join on
+        :param right_index: If true uses the index of the right df/series as columns to join on
+        :param suffixes: Tuple of two strings. Will be used to suffix duplicate column names. Must make
+            column names unique
+        :return: A new Dataframe. The original frames are not modified.
         """
         from bach.merge import merge
         return merge(
@@ -1878,7 +1828,7 @@ def dict_name_series_equals(a: Dict[str, 'Series'], b: Dict[str, 'Series']):
     )
 
 
-def _escape_parameter_characters(conn: Connection, raw_sql: str) -> str:
+def escape_parameter_characters(conn: Connection, raw_sql: str) -> str:
     """
     Return a modified copy of the given sql with the query-parameter special characters escaped.
     e.g. if the connection uses '%' to mark a parameter, then all occurrences of '%' will be replaced by '%%'
